Fourth Commit

+++ src/app/pages/career/career.component.ts
@@ -1,5 +1,11 @@
 import { StringUtils, transformSeasonString } from './../../shared/utils/utils';
-import { ChangeDetectionStrategy, Component, inject, OnDestroy, OnInit } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  inject,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
 import { HeaderComponent } from '../../shared/header/header.component';
 import { ActivatedRoute, Params, Router, RouterLink } from '@angular/router';
 import { CareerService } from '../services/career.service';
@@ -11,6 +17,7 @@ import { Subscription } from 'rxjs';
 import { ModalComponent } from '../../shared/components/modal/modal.component';
 import { NewCareer } from '../../models/career/new-career';
 import { LoaderModule } from '../../shared/components/loader/loader.module';
+import { Top3StatsPlayersCareerComponent } from '../../shared/components/top-3-stats-players-career/top-3-stats-players-career.component';
 @Component({
   selector: 'app-career',
   standalone: true,
@@ -19,11 +26,12 @@ import { LoaderModule } from '../../shared/components/loader/loader.module';
     RouterLink,
     DetailsManagerClubIconComponent,
     ModalComponent,
-    LoaderModule
+    LoaderModule,
+    Top3StatsPlayersCareerComponent,
   ],
   templateUrl: './career.component.html',
   styleUrl: './career.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CareerComponent implements OnInit, OnDestroy {
   #careerService = inject(CareerService);
@@ -36,8 +44,7 @@ export class CareerComponent implements OnInit, OnDestroy {
   public getCareerDetails = this.#careerService.getCareerDetails;
   public initialSeason = this.#careerService.getSeasonByInitialSeason;
   public seasons = this.#careerService.getAllSeasonsByCareer;
-  public showSeason = transformSeasonString
-
+  public showSeason = transformSeasonString;
 
   private confirmSubscription!: Subscription;
 
@@ -48,12 +55,13 @@ export class CareerComponent implements OnInit, OnDestroy {
         this.#careerService.httpSeasonsByCareer$(this.id).subscribe();
         this.#careerService.httpCareersById$(this.id).subscribe({
           next: (career: NewCareer) => {
-            this.#careerService.httpSeasonByInitialSeason$(career.fifaCareer).subscribe();
-          }
+            this.#careerService
+              .httpSeasonByInitialSeason$(career.fifaCareer)
+              .subscribe();
+          },
         });
       },
     });
-
   }
 
   ngOnDestroy() {
@@ -63,20 +71,18 @@ export class CareerComponent implements OnInit, OnDestroy {
   }
 
   public openModalConfirmation(): void {
-    this.subscribeToConfirm(
-      () => {
-        console.log('User confirmed the action.');
-        this.deleteCareer();
-      },
-      () => {
-        console.log('User cancelled the action.');
-      }
-    );
     this.#modalService.showConfirmation(
       'Tem certeza que deseja deletar essa carreira?',
       'Sim',
       'Não'
     );
+    this.#modalService.confirmState().subscribe((confirmed) => {
+      if (confirmed) {
+        this.deleteCareer();
+      } else {
+        return;
+      }
+    });
   }
 
   public formatSeason(fifaCareer: string): string {
@@ -97,20 +103,5 @@ export class CareerComponent implements OnInit, OnDestroy {
         });
       },
     });
-  }
-
-  private subscribeToConfirm(onConfirm: () => void, onCancel: () => void) {
-    if (this.confirmSubscription) {
-      this.confirmSubscription.unsubscribe();
-    }
-    this.confirmSubscription = this.#modalService.confirmState$.subscribe(
-      (isConfirmed) => {
-        if (isConfirmed) {
-          onConfirm();
-        } else {
-          onCancel();
-        }
-      }
-    );
   }
 }
